+++ ilpv4-connector-server/src/test/java/com/sappenin/interledger/ilpv4/connector/server/spring/settings/ConnectorSettingsFromPropertyFileTest.java
@@ -20,7 +20,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.math.BigInteger;
 import java.time.Duration;
 import java.util.List;
 import java.util.Optional;
@@ -92,7 +91,7 @@ public class ConnectorSettingsFromPropertyFileTest {
       assertThat(account.getAssetScale(), is(8));
       assertThat(account.isSendRoutes(), is(false));
       assertThat(account.isReceiveRoutes(), is(false));
-      assertThat(account.getMaximumPacketAmount(), is(Optional.of(BigInteger.valueOf(100001L))));
+      assertThat(account.getMaximumPacketAmount(), is(Optional.of(100001L)));
       assertThat(account.getCustomSettings().get("foo"), is("bar"));
       assertThat(account.getCustomSettings().get("boo"), is("baz"));
 
@@ -118,7 +117,7 @@ public class ConnectorSettingsFromPropertyFileTest {
       assertThat(account.getAssetScale(), is(3));
       assertThat(account.isSendRoutes(), is(true));
       assertThat(account.isReceiveRoutes(), is(true));
-      assertThat(account.getMaximumPacketAmount(), is(Optional.of(BigInteger.valueOf(100002L))));
+      assertThat(account.getMaximumPacketAmount(), is(Optional.of(100002L)));
       assertThat(account.getCustomSettings().get("foo"), is("bar1"));
       assertThat(account.getCustomSettings().get("boo"), is("baz1"));
 
@@ -197,7 +196,7 @@ public class ConnectorSettingsFromPropertyFileTest {
       assertThat(settings.getAssetScale(), is(8));
       assertThat(settings.isSendRoutes(), is(false));
       assertThat(settings.isReceiveRoutes(), is(false));
-      assertThat(settings.getMaximumPacketAmount(), is(Optional.of(BigInteger.valueOf(100001L))));
+      assertThat(settings.getMaximumPacketAmount(), is(Optional.of(100001L)));
       assertThat(settings.getCustomSettings().get("foo"), is("bar"));
       assertThat(settings.getCustomSettings().get("boo"), is("baz"));
